moved get_file
