+++ tf_bodypix/utils/io.py
@@ -1,0 +1,14 @@
+import os
+from hashlib import md5
+
+import tensorflow as tf
+
+
+def get_file(file_path: str) -> str:
+    if os.path.exists(file_path):
+        return file_path
+    local_path = tf.keras.utils.get_file(
+        md5(file_path.encode('utf-8')).hexdigest() + '-' + os.path.basename(file_path),
+        file_path
+    )
+    return local_path
